Split SDL input into separate files for 1.2 and 2
