+++ client/sdl/i_input_sdl12.h
@@ -20,8 +20,7 @@
 //
 //-----------------------------------------------------------------------------
 
-#ifndef __I_SDLINPUT_H__
-#define __I_SDLINPUT_H__
+#pragma once
 
 #include "i_sdl.h"
 #include "doomtype.h"
@@ -188,160 +187,3 @@ private:
 };
 
 #endif	// SDL12
-
-#ifdef SDL20
-
-// ============================================================================
-//
-// ISDL20KeyboardInputDevice class interface
-//
-// ============================================================================
-
-class ISDL20KeyboardInputDevice : public IKeyboardInputDevice
-{
-public:
-	ISDL20KeyboardInputDevice(int id);
-	virtual ~ISDL20KeyboardInputDevice();
-
-	virtual bool active() const;
-
-	virtual void pause();
-	virtual void resume();
-	virtual void reset();
-
-	virtual void gatherEvents();
-
-	virtual bool hasEvent() const
-	{	return !mEvents.empty();	}
-
-	virtual void getEvent(event_t* ev);
-
-	virtual void flushEvents();
-
-	virtual void enableTextEntry();
-	virtual void disableTextEntry();
-
-private:
-	int translateKey(SDL_Keysym keysym);
-	int getTextEventValue();
-
-	bool					mActive;
-	bool					mTextEntry;
-
-	typedef std::queue<event_t> EventQueue;
-	EventQueue				mEvents;
-};
-
-
-// ============================================================================
-//
-// ISDL20MouseInputDevice class interface
-//
-// ============================================================================
-
-class ISDL20MouseInputDevice : public IInputDevice
-{
-public:
-	ISDL20MouseInputDevice(int id);
-	virtual ~ISDL20MouseInputDevice();
-
-	virtual bool active() const;
-
-	virtual void pause();
-	virtual void resume();
-	virtual void reset();
-
-	virtual void gatherEvents();
-
-	virtual bool hasEvent() const
-	{	return !mEvents.empty();	}
-
-	virtual void getEvent(event_t* ev);
-
-	virtual void flushEvents();
-
-private:
-	bool			mActive;
-
-	typedef std::queue<event_t> EventQueue;
-	EventQueue		mEvents;
-};
-
-
-// ============================================================================
-//
-// ISDL20JoystickInputDevice class interface
-//
-// ============================================================================
-
-class ISDL20JoystickInputDevice : public IInputDevice
-{
-public:
-	ISDL20JoystickInputDevice(int id);
-	virtual ~ISDL20JoystickInputDevice();
-
-	virtual bool active() const;
-
-	virtual void pause();
-	virtual void resume();
-	virtual void reset();
-
-	virtual void gatherEvents();
-
-	virtual bool hasEvent() const
-	{	return !mEvents.empty();	}
-
-	virtual void getEvent(event_t* ev);
-
-	virtual void flushEvents();
-
-private:
-	static const int JOY_DEADZONE = 6000;
-
-	bool			mActive;
-
-	typedef std::queue<event_t> EventQueue;
-	EventQueue		mEvents;
-
-	int				mJoystickId;
-	SDL_GameController*	mJoystick;
-};
-
-
-// ============================================================================
-//
-// ISDL20InputSubsystem class interface
-//
-// ============================================================================
-
-class ISDL20InputSubsystem : public IInputSubsystem
-{
-public:
-	ISDL20InputSubsystem();
-	virtual ~ISDL20InputSubsystem();
-
-	virtual void grabInput();
-	virtual void releaseInput();
-
-	virtual bool isInputGrabbed() const
-	{	return mInputGrabbed;	}
-
-	virtual std::vector<IInputDeviceInfo> getKeyboardDevices() const;
-	virtual void initKeyboard(int id);
-	virtual void shutdownKeyboard(int id);
-
-	virtual std::vector<IInputDeviceInfo> getMouseDevices() const;
-	virtual void initMouse(int id);
-	virtual void shutdownMouse(int id);
-
-	virtual std::vector<IInputDeviceInfo> getJoystickDevices() const;
-	virtual void initJoystick(int id);
-	virtual void shutdownJoystick(int id);
-
-private:
-	bool				mInputGrabbed;
-};
-
-#endif	// SDL20
-
-#endif	// __I_SDLINPUT_H__
